requirements.txt
django==3.2.15
django-debug-toolbar==3.2.1
Pillow==8.2.0
environs[django]==9.3.2
django-phonenumber-field==8.0.0
django-phonenumbers==1.0.1
djangorestframework==3.14.0
